feat(backend): get /api/user with token
