+++ backend/exposed_api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from . import project_views, views
+from . import project_views, user_views, views
 from .event_views import EventListView
 
 app_name = "exposed_api"
@@ -12,8 +12,10 @@ urlpatterns = [
     path("projects/<int:_id>/", project_views.ProjectDetailView.as_view(), name="project_detail_or_crud"),
     # Events endpoint
     path("events/", EventListView.as_view(), name="events_list"),
+    # User endpoints
+    path("user/", user_views.get_current_user, name="current_user"),
+    path("user/<int:user_id>/", user_views.user_detail, name="user_detail"),
     # Other endpoints
-    path("user/<int:user_id>/", views.user_detail, name="user_detail"),
     path("projectViews/<int:user_id>/", views.project_views, name="project_views"),
     path("projectEngagement/<int:user_id>/", views.project_engagement, name="project_engagement"),
 ]
